mqtt-setup-support/requirements.txt
azure-iot-hub==2.6.0
azure-iot-device==2.11.0
google-cloud-core==2.3.0
google-cloud-iot==2.5.0
python-dotenv==0.20.0
msrest==0.6.21
protobuf==3.20.1
requests==2.26.0
urllib3==1.26.6
pyyaml==6.0
